+++ mockspresso-dagger/src/test/java/com/episode6/hackit/mockspresso/dagger/DaggerPluginsExtTest.kt
@@ -1,0 +1,23 @@
+package com.episode6.hackit.mockspresso.dagger
+
+import com.episode6.hackit.mockspresso.Mockspresso
+import com.nhaarman.mockitokotlin2.any
+import com.nhaarman.mockitokotlin2.mock
+import com.nhaarman.mockitokotlin2.verify
+import org.junit.Test
+import org.mockito.stubbing.Answer
+
+/**
+ * Tests [com.episode6.hackit.mockspresso.dagger.DaggerPluginsExtKt]
+ */
+class DaggerPluginsExtTest {
+
+  val builder: Mockspresso.Builder = mock(defaultAnswer = Answer { it.mock })
+
+  @Test
+  fun testSimplePluginSourceOfTruth() {
+    builder.injectByDaggerConfig()
+
+    verify(builder).plugin(any<DaggerMockspressoPlugin>())
+  }
+}