+++ CHANGELOG.md
@@ -16,6 +16,7 @@
  - Added kotlin convenience extension methods for built in plugins
      - `Builder.injectBySimpleConfig()`: Applies the simple injection configuration plugin
      - `Builder.injectByJavaxConfig()`: Applies the Javax injection configuration plugin
+     - `Builder.injectByDaggerConfig()`: Applies the dagger injection configuration plugin
      - `Builder.mockByMockito()`: Applies the mockito mocker config
      - `Builder.automaticFactories(vararg KClass<*>)`: Special object handling using `MockitoAutoFactoryMaker`